fixing phone number validation

+++ app/src/main/java/com/tcd/yaatra/services/api/yaatra/models/RegisterRequestBody.java
@@ -27,7 +27,7 @@ public class RegisterRequestBody {
     @SerializedName("country")
     private String country;
     @SerializedName("phone_number")
-    private int phoneNumber;
+    private long phoneNumber;
 
     public RegisterRequestBody() {
     }
@@ -121,10 +121,10 @@ public class RegisterRequestBody {
         this.country = country;
     }
 
-    public int getPhoneNumber() {
+    public long getPhoneNumber() {
         return phoneNumber;
     }
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
